MuktiBhai_Langgraph_Bot/requirements.txt
langgraph
python-dotenv
langchain_community
langchain
langchain-google-genai
langchain_openai
pinecone[grpc]
fastapi
uvicorn